fix(loader): lifecycle loader support named export
