+++ test/fixtures/app_with_lifecycle/bootstrap_load.ts
@@ -8,7 +8,7 @@ import LifecycleList from './lifecyclelist';
 export const TEST_LIFE_CYCLE_LIST = 'TEST_LIFE_CYCLE_LIST';
 
 @LifecycleHookUnit()
-export default class AppLifecycle implements ApplicationLifecycle {
+export default class AppLoadLifecycle implements ApplicationLifecycle {
   @Inject(ArtusInjectEnum.Application)
   app: ArtusApplication;
 
@@ -35,23 +35,5 @@ export default class AppLifecycle implements ApplicationLifecycle {
   async  didLoad() {
     await new Promise(resolve => setTimeout(resolve, 100));
     this.lifecycleList.add('app_didLoad');
-  }
-
-  @LifecycleHook('willReady')
-  async willReady() {
-    await new Promise(resolve => setTimeout(resolve, 100));
-    this.lifecycleList.add('app_willReady');
-  }
-
-  @LifecycleHook('didReady')
-  async didReady() {
-    await new Promise(resolve => setTimeout(resolve, 100));
-    this.lifecycleList.add('app_didReady');
-  }
-
-  @LifecycleHook()
-  async beforeClose() {
-    await new Promise(resolve => setTimeout(resolve, 100));
-    this.lifecycleList.add('app_beforeClose');
   }
 }
